feat: move to scss files and components clinet dir
