+++ client/src/components/ui/Main/TrainFrame.tsx
