+++ client/src/styles/components/ui/Main/aboutFrame.module.scss
@@ -1,5 +1,5 @@
-@use '../../font' as *;
-@use '../../mediaQuery' as *;
+@use '../../../font' as *;
+@use '../../../mediaQuery' as *;
 
 #debug, #effect, #aboutFrame, #group{
     display: flex;
