+++ client/src/styles/components/ui/Main/modalFrame.module.scss
@@ -1,4 +1,4 @@
-@use '../../font' as *;
+@use '../../../font' as *;
 
 .debug{
     display: flex;
