+++ client/src/styles/components/ui/Main/roadMapFrame.module.scss
@@ -1,5 +1,5 @@
-@use '../../font' as *;
-@use '../../mediaQuery' as *;
+@use '../../../font' as *;
+@use '../../../mediaQuery' as *;
 
 #debug{
     display: flex;
