+++ client/src/styles/components/ui/Main/trainFrame.module.scss
@@ -1,4 +1,4 @@
-@use '../../mediaQuery' as *;
+@use '../../../mediaQuery' as *;
 
 #debug{
     display: flex;
