+++ client/src/styles/components/ui/Main/translate.module.scss
@@ -1,4 +1,4 @@
-@use '../../mediaQuery' as *;
+@use '../../../mediaQuery' as *;
 
 .debug{
     position: relative;
